fix(engine): add the parent seals if no seals were gossipped to us
note:
The optimal approach here would be to use the union of the already
existing parent seals and the ones received during the gossipping stage
(instead of using either one or the other)

The hard part about this is that the parent's committed seals are already aggregated, meaning we cannot meaningfully get the union of an already aggregated signature with the ones we received from the gossipping.

Also, there is no good way for us to fetch the signatures before they
were aggregated, since we only store the aggregated signature and not
the BLS signatures before their aggregation.